+++ frontend/src/pages/Statistics.tsx
@@ -1,8 +1,5 @@
 import { type FC } from "react";
 import {
-  PieChart,
-  Pie,
-  Cell,
   ResponsiveContainer,
   Tooltip,
   LineChart,
@@ -13,10 +10,7 @@ import {
 } from "recharts";
 import { useAppSelector } from "@/store/store";
 import { statsAPI } from "@/api/statsAPI";
-import type { UserStatisticsResponse } from "@/types/api.dto";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
   Download,
@@ -30,8 +24,8 @@ import {
   TrendingUp,
   Layers,
 } from "lucide-react";
-
-const PIE_COLORS = ["#3b82f6", "#22c55e", "#f59e0b", "#ef4444"];
+import { StatisticsSkeleton } from "@/components/statistics/Skeleton";
+import { PieChartCard } from "@/components/statistics/PieChart";
 
 const projectCards = [
   { key: "species", icon: Bug, label: "Видов", field: "species_count" as const },
@@ -80,79 +74,6 @@ const commonLabels = [
 
 function formatNumber(n: number): string {
   return n.toLocaleString("ru-RU");
-}
-
-function PieChartCard({
-  data,
-  error,
-}: {
-  data: UserStatisticsResponse | undefined;
-  error: boolean;
-}) {
-  if (error) {
-    return (
-      <Card>
-        <CardContent className="py-6 text-center text-sm text-red-500">
-          Не удалось загрузить статистику
-        </CardContent>
-      </Card>
-    );
-  }
-  if (!data || data.top_species.length === 0) {
-    return (
-      <Card>
-        <CardContent className="py-12 text-center text-sm text-muted-foreground">
-          У вас пока нет записей
-        </CardContent>
-      </Card>
-    );
-  }
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle className="flex items-center gap-2 text-sm font-medium text-muted-foreground">
-          <Bug className="size-4" />
-          Ваши топ-виды
-        </CardTitle>
-      </CardHeader>
-      <CardContent>
-        <div className="flex flex-col items-center gap-6 sm:flex-row sm:items-start">
-          <ResponsiveContainer width="100%" height={240} className="max-w-xs shrink-0">
-            <PieChart>
-              <Pie
-                data={data.top_species}
-                dataKey="count"
-                nameKey="species"
-                cx="50%"
-                cy="50%"
-                outerRadius={100}
-                innerRadius={50}
-              >
-                {data.top_species.map((_, i) => (
-                  <Cell key={i} fill={PIE_COLORS[i % PIE_COLORS.length]} />
-                ))}
-              </Pie>
-              <Tooltip formatter={(value, name) => [formatNumber(Number(value)), String(name)]} />
-            </PieChart>
-          </ResponsiveContainer>
-          <div className="w-full space-y-2 sm:w-auto">
-            {data.top_species.map((item, i) => (
-              <div key={item.species} className="flex items-center gap-3 text-sm">
-                <span
-                  className="inline-block size-3 shrink-0 rounded-full"
-                  style={{ backgroundColor: PIE_COLORS[i % PIE_COLORS.length] }}
-                />
-                <span className="truncate">{item.species}</span>
-                <Badge variant="secondary" className="ml-auto shrink-0 font-mono text-xs">
-                  {formatNumber(item.count)}
-                </Badge>
-              </div>
-            ))}
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
 }
 
 function Value({ value }: { value: number | string | null | undefined }) {
@@ -223,42 +144,7 @@ const Statistics: FC = () => {
   const isLoading = projectLoading || userLoading;
 
   if (isLoading) {
-    return (
-      <div className="space-y-6">
-        <Skeleton className="h-7 w-32" />
-        <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
-          {Array.from({ length: 4 }).map((_, i) => (
-            <Card key={i}>
-              <CardHeader>
-                <Skeleton className="h-4 w-24" />
-              </CardHeader>
-              <CardContent>
-                <Skeleton className="h-8 w-16" />
-              </CardContent>
-            </Card>
-          ))}
-        </div>
-        <Skeleton className="h-32 w-full rounded-xl" />
-        <div className="flex gap-4">
-          {Array.from({ length: 3 }).map((_, i) => (
-            <Card key={i} className="flex-1">
-              <CardContent className="pt-4">
-                <Skeleton className="mb-2 h-4 w-20" />
-                <Skeleton className="h-5 w-32" />
-              </CardContent>
-            </Card>
-          ))}
-        </div>
-        <Card>
-          <CardHeader>
-            <Skeleton className="h-5 w-32" />
-          </CardHeader>
-          <CardContent className="flex justify-center">
-            <Skeleton className="size-48 rounded-full" />
-          </CardContent>
-        </Card>
-      </div>
-    );
+    return <StatisticsSkeleton />;
   }
 
   if (projectError) {
